fix Go lint for model catalog and allocator
